requirements.txt
fastapi
uvicorn[standard]
websockets
faster-whisper
pyaudio
soundfile
numpy
torch
nvidia-cublas-cu12
nvidia-cudnn-cu12
openai
ruff
